+++ CHANGELOG.md
@@ -2900,3 +2900,4 @@
 - [2026-07-16 16:01] style(i18n): harmonize light and dark theme palette for i18n
 - [2026-07-16 16:37] feat(state): add bi-directional event bus handler for state
 - [2026-07-16 17:12] fix(cache): resolve race condition during cache initialization
+- [2026-07-16 17:28] fix(perf): resolve race condition during perf initialization